+++ app/src/main/java/org/kiwix/kiwixmobile/data/local/KiwixDatabase.java
@@ -17,7 +17,7 @@
  * MA 02110-1301, USA.
  */
 
-package org.kiwix.kiwixmobile.database;
+package org.kiwix.kiwixmobile.data.local;
 
 import android.content.Context;
 import android.util.Log;
@@ -27,11 +27,12 @@ import com.yahoo.squidb.data.adapter.SQLiteDatabaseWrapper;
 import com.yahoo.squidb.sql.Table;
 
 import org.kiwix.kiwixmobile.ZimContentProvider;
-import org.kiwix.kiwixmobile.database.entity.BookDatabaseEntity;
-import org.kiwix.kiwixmobile.database.entity.Bookmarks;
-import org.kiwix.kiwixmobile.database.entity.LibraryDatabaseEntity;
-import org.kiwix.kiwixmobile.database.entity.NetworkLanguageDatabaseEntity;
-import org.kiwix.kiwixmobile.database.entity.RecentSearch;
+import org.kiwix.kiwixmobile.data.local.dao.BookmarksDao;
+import org.kiwix.kiwixmobile.data.local.entity.BookDatabaseEntity;
+import org.kiwix.kiwixmobile.data.local.entity.Bookmarks;
+import org.kiwix.kiwixmobile.data.local.entity.LibraryDatabaseEntity;
+import org.kiwix.kiwixmobile.data.local.entity.NetworkLanguageDatabaseEntity;
+import org.kiwix.kiwixmobile.data.local.entity.RecentSearch;
 
 import java.io.BufferedReader;
 import java.io.FileNotFoundException;
