+++ app/src/main/java/org/kiwix/kiwixmobile/data/local/dao/BookmarksDao.java
@@ -15,13 +15,14 @@
  * You should have received a copy of the GNU General Public License
  * along with this program. If not, see <http://www.gnu.org/licenses/>.
  */
-package org.kiwix.kiwixmobile.database;
+package org.kiwix.kiwixmobile.data.local.dao;
 
 
 import com.yahoo.squidb.data.SquidCursor;
 import com.yahoo.squidb.sql.Query;
 
-import org.kiwix.kiwixmobile.database.entity.Bookmarks;
+import org.kiwix.kiwixmobile.data.local.KiwixDatabase;
+import org.kiwix.kiwixmobile.data.local.entity.Bookmarks;
 
 import java.util.ArrayList;
