+++ app/src/main/java/org/kiwix/kiwixmobile/data/local/entity/RecentSearchSpec.java
@@ -15,21 +15,22 @@
  * You should have received a copy of the GNU General Public License
  * along with this program. If not, see <http://www.gnu.org/licenses/>.
  */
-package org.kiwix.kiwixmobile.database.entity;
+package org.kiwix.kiwixmobile.data.local.entity;
 
 import com.yahoo.squidb.annotations.ColumnSpec;
 import com.yahoo.squidb.annotations.TableModelSpec;
 
 /**
- * Squidb spec for saved bookmarks.
+ * Squidb spec for recent searches.
  */
-@TableModelSpec(className = "Bookmarks", tableName = "Bookmarks")
-public class BookmarksSpec {
+@TableModelSpec(className = "RecentSearch", tableName = "recentSearches")
+public class RecentSearchSpec {
 
   @ColumnSpec(constraints = "NOT NULL")
-  public String ZimId;
-  public String ZimName;
-  public String bookmarkUrl;
-  public String bookmarkTitle;
+  public String searchString;
+
+  @ColumnSpec(constraints = "NOT NULL")
+  public String zimID;
+
 
 }
