+++ app/src/main/java/org/kiwix/kiwixmobile/search/SearchPresenter.java
@@ -18,7 +18,7 @@
 package org.kiwix.kiwixmobile.search;
 
 import org.kiwix.kiwixmobile.base.BasePresenter;
-import org.kiwix.kiwixmobile.database.RecentSearchDao;
+import org.kiwix.kiwixmobile.data.local.dao.RecentSearchDao;
 
 import javax.inject.Inject;
